+++ src/Common.jl
@@ -72,24 +72,15 @@ sortDFG(vars::Vector{Symbol}; lt = natural_lt, kwargs...) = sort(vars; lt = lt, 
 ##==============================================================================
 ## Validation of session, robot, and user labels.
 ##==============================================================================
-global _invalidIds = [
-    "USER",
-    "ROBOT",
-    "SESSION",
-    "AGENT",
-    "VARIABLE",
-    "FACTOR",
-    "PPE",
-    "BLOB_ENTRY",
-    "FACTORGRAPH",
-]
+global _invalidIds =
+    ["GRAPH", "AGENT", "VARIABLE", "FACTOR", "PPE", "BLOB_ENTRY", "FACTORGRAPH"]
 
 const global _validLabelRegex::Regex = r"^[a-zA-Z][-\w\.\@]*$"
 
 """
 $(SIGNATURES)
 
-Returns true if the label is valid for a session, robot, or user ID.
+Returns true if the label is valid for node.
 """
 function isValidLabel(id::Union{Symbol, String})
     if typeof(id) == Symbol